Refactor email service import paths to use clients directory
