+++ src/app.module.ts
@@ -16,7 +16,7 @@ import { IsUniqueEmailConstraint } from "src/common/validators/unique-email.vali
 import { ApiKeyGuard } from "src/middlewares/api-key.middleware";
 import { ApiKeysController } from './services/api-keys/api-keys.controller';
 import { ApiKeysService } from './services/api-keys/api-keys.service';
-import { EmailService } from 'src/services/email/email.service';
+import { EmailService } from 'src/clients/email/email.service';
 
 @Module({
 imports: [ConfigModule.forRoot({
